fix(deps): Recover build structure (Closes #18)
With latest dependencies update I have included the spec files and mock into typechecking.
This eventually changes the build structure emitted by Typescript, and hence made all the package imports
unavailable.

With this fix we create a build tsconfig, which excludes spec files, so recovers the old dist
output. In the meantime in development we still typecheck spec files.